Move prebuild helper from workflow view to workflow manager
